Client: Renamed post-page directory, component file, and styles file to create-post. Created new post-page directory, component file, and styles file and started implementing new post-page component for view a single post in a page with path '/p/:postId'
