+++ client/src/pages/post-page/post-page.styles.scss
@@ -1,56 +1,10 @@
-.homepage {
+.post-page {
   min-height: 800px;
-  padding: 40px 0;
-}
-
-.upload {
-  margin-top: 20px;
   display: flex;
+  width: 85vw;
   flex-direction: column;
   align-items: center;
-}
-
-.img-preview-container {
-  margin-top: 20px;
-  height: 400px;
-  width: 400px;
-  border: solid 3px;
-
-  .img-preview-placeholder {
-    position: relative;
-    font-size: 20px;
-    margin-top: 50%;
-    margin-left: 50%;
-    transform: translate(-50%, -50%);
-    width: 300px;
-    height: 300px;
-    border: dashed 2px;
-
-    .placeholder-text-container {
-      position: relative;
-      margin-top: 50%;
-      margin-left: 50%;
-      transform: translate(-50%, -50%);
-
-      .placeholder-text {
-        font-weight: 420;
-      }
-    }
-  }
-
-  .alert {
-    position: relative;
-    width: 300px;
-    margin-top: 50%;
-    margin-left: 50%;
-    transform: translate(-50%, -65%);
-  }
-
-  .img-preview {
-    margin-top: 50%;
-    transform: translateY(-50%);
-    max-width: 300px;
-    max-height: 300px;
-    height: auto;
-  }
+  margin-left: 50%;
+  margin-top: 40px;
+  transform: translate(-50%);
 }
